App leftbar and Server Leftbar were advanced.
On this commit the app leftbar and server leftbar were advanced.
Several business logic was added to the project. Invite links were added
as a feature of servers in EDEN. Edit server and Manage members of the
servers were also added but there aren't in their final stages. However,
they were advanced in a good way on this day. Also, logic for adding
people to servers and providers for better performance were added.